bulk_analyzer/requirements.txt
fuzzywuzzy
python-levenshtein
gitpython
pyyaml
requests
matplotlib
pandas
seaborn
numpy
kneed
notebook
urllib3